% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_matrix_model.r
\name{ModelMatrixModel}
\alias{ModelMatrixModel}
\title{ModelMatrixModel() function}
\usage{
ModelMatrixModel(
  rformula,
  data,
  sparse = T,
  center = FALSE,
  scale = FALSE,
  remove_1st_dummy = F,
  remove_same_value_column = F,
  verbose = T
)
}
\arguments{
\item{rformula}{a formula, e.g. output of formula()}

\item{data}{a data.frame}

\item{sparse}{bool, if True return a sparse matrix.}

\item{center}{bool, if center the output}

\item{scale}{bool, if scale the output}

\item{remove_1st_dummy}{bool, if remove the first dummy variable in one hot key transformation}

\item{remove_same_value_column}{bool, if remove column with one unique value}

\item{verbose}{bool, if print out progress}
}
\value{
A ModelMatrixModel class,which includes the transformed matrix and  the fitted parameters.
}
\description{
This function create model.matrix and save the fitted parameters
}
