% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_matrix_model.r
\name{predict.ModelMatrixModel}
\alias{predict.ModelMatrixModel}
\title{Create model.matrix and save the fitted parameters}
\usage{
\method{predict}{ModelMatrixModel}(object, data, handleInvalid = "keep", ...)
}
\arguments{
\item{object}{a ModelMatrixModel object}

\item{data}{a data.frame}

\item{handleInvalid}{a string,'keep' or 'error'.  In dummy variable transformation, if categorical variable has a factor level that is unseen before, 'keep' will keep the record, output dummy variables will be all zero.}

\item{...}{other parameters}
}
\value{
A ModelMatrixModel class,which includes the transformed matrix and  the fitted parameters copied from input object.
}
\description{
This function transform new data based on information from a ModelMatrixModel object
}
